UML-Project
Keep in mind gents this was done in a nights time there are bugs and I
forgot about it but there is also a glaring issue with the creation of
the "class" shape and the text that goes along with it.  I will focus on
cleaning it up a bit tomorrow/over the weekend.

As usual any and all suggestions/complaints are welcomed and wanted,
voice your thoughts! Its our project.