Line Class
This is from the previous version from where you cannot move the classes
around. I will update it and upload that one tomorrow. As of now i know
of a couple bugs, such as not being able to draw more than one line, and
i will work on that tomorrow too. Sorry it took me so long, this is my
first time using github, but here's everything